refactor(kai): extract market discovery components
